requirements.txt
PyQt5==5.15.10
Pillow==10.0.1
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
bcrypt==4.0.1cl